requirements.txt
colour-science==0.4.4
contourpy==1.2.0
cycler==0.12.1
fonttools==4.49.0
imageio==2.34.0
kiwisolver==1.4.5
matplotlib==3.8.3
numpy==1.26.4
packaging==23.2
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.9.0.post0
scipy==1.12.0
six==1.16.0
typing_extensions==4.10.0
